+++ src/tracer/src/daemon/daemon_run.rs
@@ -3,8 +3,9 @@ use crate::client::exporters::log_forward::LogForward;
 use crate::client::exporters::log_writer::LogWriterEnum;
 use crate::client::TracerClient;
 use crate::common::types::cli::params::FinalizedInitArgs;
+use crate::config::Config;
 use crate::daemon::server::DaemonServer;
-use crate::{config::Config, utils::emit_analytic_event};
+use crate::utils::analytics::emit_analytic_event;
 use anyhow::{Context, Result};
 use std::net::SocketAddr;
 use tracing::info;
