+++ src/tracer/src/utils/system_info.rs
@@ -1,0 +1,34 @@
+use std::process::Command;
+
+pub fn check_sudo_privileges() {
+    if std::env::var("SUDO_USER").is_err() {
+        println!("Warning: Running without sudo privileges. Some operations may fail.");
+        // Get the current executable path and arguments
+        let current_exe =
+            std::env::current_exe().unwrap_or_else(|_| std::path::PathBuf::from("tracer"));
+        let args: Vec<String> = std::env::args().collect();
+        let sudo_command = format!("sudo {} {}", current_exe.display(), args[1..].join(" "));
+        println!("Try running with elevated privileges:\n {}", sudo_command);
+    }
+}
+
+pub fn get_kernel_version() -> Option<(u32, u32)> {
+    let kernel_version = Command::new("uname")
+        .arg("-r")
+        .output()
+        .ok()
+        .and_then(|output| {
+            String::from_utf8(output.stdout).ok().and_then(|version| {
+                let parts: Vec<&str> = version.trim().split('.').collect();
+                if parts.len() >= 2 {
+                    let major = parts[0].parse::<u32>().ok()?;
+                    let minor = parts[1].parse::<u32>().ok()?;
+                    Some((major, minor))
+                } else {
+                    None
+                }
+            })
+        });
+
+    kernel_version
+}
